Lift risk types to domain
