+++ packages/app/src/features/dialogs/borrow/views/BorrowView.tsx
@@ -1,6 +1,3 @@
-import BigNumber from 'bignumber.js'
-import { UseFormReturn } from 'react-hook-form'
-
 import { TokenWithBalance, TokenWithValue } from '@/domain/common/types'
 import { Objective } from '@/features/actions/logic/types'
 import { DialogActionsPanel } from '@/features/dialogs/common/components/DialogActionsPanel'
@@ -10,7 +7,8 @@ import { DialogForm } from '@/features/dialogs/common/components/form/DialogForm
 import { AssetInputSchema } from '@/features/dialogs/common/logic/form'
 import { FormFieldsForDialog, PageStatus } from '@/features/dialogs/common/types'
 import { DialogTitle } from '@/ui/atoms/dialog/Dialog'
-
+import BigNumber from 'bignumber.js'
+import { UseFormReturn } from 'react-hook-form'
 import { BorrowOverviewPanel } from '../components/BorrowOverviewPanel'
 
 export interface BorrowViewProps {
